ensure plantuml invoke can print list-based cmd options
When a project is configured with a list-based command:

 plantuml = ['java', '-jar', f'{plantuml_jar}']

And a platform cannot run (e.g. no Java installed), the error scenario
results in a `TypeError` exception. For example:

 File "...\.tox\py\Lib\site-packages\sphinxcontrib\plantuml.py", line 397, in render
     'plantuml command %r cannot be run' % self.builder.config.plantuml
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 TypeError: not all arguments converted during string formatting

Adjusting the error message to use an f-string allows both string and
list typed values to be added into the string. For example:

 sphinxcontrib.plantuml.PlantUmlError: plantuml command ('java', '-jar', '...\plantuml\\plantuml.jar') cannot be run

Signed-off-by: James Knight <git@jdknight.me>